Adding additional REST V2 tests

+++ redback-integrations/redback-rest/redback-rest-services/src/main/java/org/apache/archiva/redback/rest/services/v2/DefaultGroupService.java
@@ -88,7 +88,7 @@ public class DefaultGroupService
     }
 
     @Override
-    public PagedResult<List<Group>> getGroups( Integer offset, Integer limit ) throws RedbackServiceException
+    public PagedResult<List<Group>> getGroups( Long offset, Long limit ) throws RedbackServiceException
     {
         LdapConnection ldapConnection = null;
